+++ src/main/java/com/hxqh/eam/model/view/VTotalMonth.java
@@ -1,8 +1,10 @@
 package com.hxqh.eam.model.view;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.io.Serializable;
-import javax.persistence.*;
-import java.math.BigDecimal;
 
 
 /**
@@ -15,43 +17,43 @@ public class VTotalMonth implements Serializable {
 	private static final long serialVersionUID = 1L;
 
 	@Column(name="W1_ANSWER")
-	private BigDecimal w1Answer;
+	private Long w1Answer;
 
 	@Column(name="W1_ATTEMPT")
-	private BigDecimal w1Attempt;
+	private Long w1Attempt;
 
 	@Id
 	@Column(name="W1_MM")
 	private String w1Mm;
 
 	@Column(name="W1_OGLOST")
-	private BigDecimal w1Oglost;
+	private Long w1Oglost;
 
 	@Column(name="W2_ANSWER")
-	private BigDecimal w2Answer;
+	private Long w2Answer;
 
 	@Column(name="W2_ATTEMPT")
-	private BigDecimal w2Attempt;
+	private Long w2Attempt;
 
 	@Column(name="W2_OGLOST")
-	private BigDecimal w2Oglost;
+	private Long w2Oglost;
 
 	public VTotalMonth() {
 	}
 
-	public BigDecimal getW1Answer() {
+	public Long getW1Answer() {
 		return this.w1Answer;
 	}
 
-	public void setW1Answer(BigDecimal w1Answer) {
+	public void setW1Answer(Long w1Answer) {
 		this.w1Answer = w1Answer;
 	}
 
-	public BigDecimal getW1Attempt() {
+	public Long getW1Attempt() {
 		return this.w1Attempt;
 	}
 
-	public void setW1Attempt(BigDecimal w1Attempt) {
+	public void setW1Attempt(Long w1Attempt) {
 		this.w1Attempt = w1Attempt;
 	}
 
@@ -63,35 +65,35 @@ public class VTotalMonth implements Serializable {
 		this.w1Mm = w1Mm;
 	}
 
-	public BigDecimal getW1Oglost() {
+	public Long getW1Oglost() {
 		return this.w1Oglost;
 	}
 
-	public void setW1Oglost(BigDecimal w1Oglost) {
+	public void setW1Oglost(Long w1Oglost) {
 		this.w1Oglost = w1Oglost;
 	}
 
-	public BigDecimal getW2Answer() {
+	public Long getW2Answer() {
 		return this.w2Answer;
 	}
 
-	public void setW2Answer(BigDecimal w2Answer) {
+	public void setW2Answer(Long w2Answer) {
 		this.w2Answer = w2Answer;
 	}
 
-	public BigDecimal getW2Attempt() {
+	public Long getW2Attempt() {
 		return this.w2Attempt;
 	}
 
-	public void setW2Attempt(BigDecimal w2Attempt) {
+	public void setW2Attempt(Long w2Attempt) {
 		this.w2Attempt = w2Attempt;
 	}
 
-	public BigDecimal getW2Oglost() {
+	public Long getW2Oglost() {
 		return this.w2Oglost;
 	}
 
-	public void setW2Oglost(BigDecimal w2Oglost) {
+	public void setW2Oglost(Long w2Oglost) {
 		this.w2Oglost = w2Oglost;
 	}
